separate middleware tests into different files
