+++ packages/datadog-plugin-express/test/middleware-disabled-test/middleware-disabled-plugin.spec.js
@@ -1,0 +1,5 @@
+'use strict'
+
+const test = require('./middleware-disabled-tests')
+
+test('plugin', [{ middleware: false }, { client: false }], {})
